no error

+++ src/log.cpp
@@ -16,7 +16,7 @@ void Log::Appender::write(const plog::Record& record) {
     if (record.getSeverity() <= plog::Severity::error)
         cmd = true;
 
-    list.push_back(Message{plog::FuncMessageFormatter::format(record), record.getSeverity(), record.getTime(), list.size() });
+    list.push_back(Message{plog::FuncMessageFormatter::format(record), record.getSeverity(), record.getTime(), (int)list.size() });
 
     // post first line
     std::ifstream ifile(File::loc()+"assets/logs/logs.txt");
